Move #generate_docset
