+++ guides/rails_guides/generator.rb
@@ -118,15 +118,6 @@ module RailsGuides
       puts "(kindlegen log at #{out})."
     end
 
-    def generate_docset
-      require 'rails_guides/dash'
-      out = "#{output_dir}/docset.out"
-      Dash.generate @source_dir, output_dir,
-        "ruby_on_rails_guides_#@version%s.docset" % (@lang.present? ? ".#@lang" : ''),
-        out
-      puts "(docset generate log at #{out})."
-    end
-
     def mobi
       "ruby_on_rails_guides_#@version%s.mobi" % (@lang.present? ? ".#@lang" : '')
     end
