+++ guides/rails_guides/generator_ja.rb
@@ -17,5 +17,14 @@ module RailsGuides
     def dash?
       @dash
     end
+
+    def generate_docset
+      require 'rails_guides/dash'
+      out = "#{output_dir}/docset.out"
+      Dash.generate @source_dir, output_dir,
+                    "ruby_on_rails_guides_#@version%s.docset" % (@lang.present? ? ".#@lang" : ''),
+                    out
+      puts "(docset generate log at #{out})."
+    end
   end
 end
